+++ product-service/handler.js
@@ -1,4 +1,4 @@
-import * as handlers from "./src";
+import * as handlers from "./src/handlers";
 
 export const getAllProducts = handlers.getAllProductsHandler;
 export const getProductById = handlers.getProductByIdHandler;
