+++ product-service/src/handlers/getProductById.js
@@ -1,5 +1,5 @@
-import { availableProducts } from "./mock/data";
-import { formatJSONResponse } from "./utils";
+import { availableProducts } from "../mock/data";
+import { formatJSONResponse } from "../utils";
 
 export const getProductByIdHandler = async (event) => {
   const { productId } = event.pathParameters || {};
