+++ product-service/src/handlers/index.js
